kvm tools: Use threading for virtio block devices
Add I/O thread to handle I/O operations in virtio-blk.

There is currently support for multiple virtio queues but the kernel side
supports only one virtio queue. It's not too much of a performance impact and
the ABI does support multiple queues there - So I've prefered to do it like
that to keep it flexible.

I/O performance itself doesn't increase much due to the patch, what changes is
system responsiveness during I/O operations.  On an unthreaded system, The VCPU
is frozen up until the I/O request is complete. On the other hand, On a
threaded system the VCPU is free to do other work or queue more I/O while
waiting for the original I/O request to complete.

[ penberg@kernel.org: cleanups ]
Signed-off-by: Sasha Levin <levinsasha928@gmail.com>
Signed-off-by: Pekka Enberg <penberg@kernel.org>