kvmtool: Remove references to top level kernel directory
The Makefile used to reference ../.. on some occasions to point
to the Linux source tree. Since this is no longer there, remove
those bogus paths.
Since we cannot (and don't want to) determine the Linux kernel
version on this way, we hardcode v3.18.0 for now, which is the version
the in-kernel-tree version used at the point of forking.
This should be later revisited and adjusted to whatever versioning
scheme we will use.

Signed-off-by: Andre Przywara <andre.przywara@arm.com>
Signed-off-by: Will Deacon <will.deacon@arm.com>