kvmtool: Copy Linux' up-to-date virtio headers
Some distributions with an older uapi header copy do not have all the
definitions that we need (Ubuntu 12.04 comes to mind).
So copy the latest versions from the Linux tree to override the system
header versions.
Also include virtio_mmio.h, which is technically not an uapi header,
but is used in such a way. QEMU has actually a copy of these
definitions in its own header file, so our approach should even be
better.
Those files are from a Linux v4.1-rc6 source tree, but the virtio-1.0
updates have been left out. kvmtool implements only the older version
of virtio.

Signed-off-by: Andre Przywara <andre.przywara@arm.com>
Signed-off-by: Will Deacon <will.deacon@arm.com>