kvmtool: Makefile: fix bogus definition of try-cc
In config/utilities.mak we have two macros checking either
compile-ability or build-ability of a library. Actually they are
defined the same, with try-cc behaving like try-build for no reason.
Fix try-cc to really only test compilation (no linking) and replace
all calls to try-cc in the Makefile with try-build (as this is what
we need). Simplify the Makefile calls a bit on the way.

Signed-off-by: Andre Przywara <andre.przywara@arm.com>
Signed-off-by: Will Deacon <will.deacon@arm.com>