Remove visible dependency files
After building, there is a lot of clutter from the dependency system.
Let's clean this up by using dir/.file.d style dependencies, similar
to those used in the Linux kernel.

In order to support this, rearrange the dependency generation to
create the dependency files as we build rather than as a separate
step, and have make clean remove them.

Signed-off-by: Russell King <rmk@arm.linux.org.uk>
[will: added make target for builtin-help.static.o]
Signed-off-by: Will Deacon <will.deacon@arm.com>