kvmtool: Switch serial input to raw mode
As I was trying to install a new VM using the Debian installer,
I noticed that the return key would work just fine in a shell,
but wouldn't do anything in the menu. Pretty annoying.

Further investigation showed that the terminal was left in
cooked mode, converting CR to LF, and thus giving the VM
the wrong information.

Clearing the ICRNL flag in the input flag set fixes the issue.

Suggested-by: Dave martin <dave.martin@arm.com>
Signed-off-by: Marc Zyngier <marc.zyngier@arm.com>
Signed-off-by: Will Deacon <will.deacon@arm.com>