Makefile: properly express guest_init dependency
So far the generation of the guest_init binaries is not properly
modelled in the Makefile: the intermediate object files are not targets.
This leads to failures when those files get deleted.
So (also in preperation for the upcoming rework) rework the dependency
chain to have those intermediate files covered as well, which involves
splitting the generation into two steps.
On the way use automatic variables where applicable and remove the
explicit listing of the guest_init targets, which are now covered by
the final $(GUEST_OBJS) targets.

Signed-off-by: Andre Przywara <andre.przywara@arm.com>
Signed-off-by: Will Deacon <will.deacon@arm.com>